% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCode.R
\name{apat}
\alias{apat}
\title{A + t(A)}
\usage{
apat(A)
}
\arguments{
\item{A}{numeric matrix}
}
\value{
numeric matrix (A + t(A))
}
\description{
A + t(A)
}
\details{
Not exported; exists for testing C code.
}
