% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPvals.R
\name{calcPvals}
\alias{calcPvals}
\title{Calculate t-SNE p-values based on a distance matrix}
\usage{
calcPvals(D, perplexity = 30)
}
\arguments{
\item{D}{numeric matrix, distance matrix}

\item{perplexity}{numeric of length 1, t-SNE perplexity}
}
\value{
numeric matrix of p-values based on the given perplexity
}
\description{
Calculate t-SNE p-values based on a distance matrix
}
\details{
Not exported; exists for testing C code.
}
