% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad.R
\name{grad}
\alias{grad}
\title{Calculate t-SNE gradient}
\usage{
grad(Y, pval, Z)
}
\arguments{
\item{Y}{numeric matrix, lower dimension embedding}

\item{pval}{numeric matrix, input data p-values}

\item{Z}{numeric covariate matrix}
}
\value{
numeric matrix, t-SNE gradient
}
\description{
Calculate t-SNE gradient
}
\details{
Not exported; exists for testing C code.
}
