% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCode.R
\name{ols}
\alias{ols}
\title{Ordinary least squares, solves B = AX for X.}
\usage{
ols(A, B)
}
\arguments{
\item{A}{numeric matrix}

\item{B}{numeric matrix}
}
\value{
numeric matrix (X)
}
\description{
Ordinary least squares, solves B = AX for X.
}
\details{
Not exported; exists for testing C code.
}
