% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqdist.R
\name{sqdist}
\alias{sqdist}
\title{Calculate squared Euclidean distance}
\usage{
sqdist(X)
}
\arguments{
\item{X}{numeric matrix}
}
\value{
numeric squared distance matrix
}
\description{
Calculate squared Euclidean distance
}
\details{
Not exported; exists for testing C code.
}
