% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssx.R
\name{ssx}
\alias{ssx}
\title{Sum of squares}
\usage{
ssx(X)
}
\arguments{
\item{X}{numeric matrix}
}
\value{
vector with the row sum of squares
}
\description{
Sum of squares
}
\details{
Not exported; exists for testing C code.
}
