% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCode.R
\name{zeroMean}
\alias{zeroMean}
\title{Subtract the column means from X}
\usage{
zeroMean(X)
}
\arguments{
\item{X}{numeric matrix}
}
\value{
numeric matrix with column means subtracted
}
\description{
Subtract the column means from X
}
\details{
Not exported; exists for testing C code.
}
